requirements.txt
django
django-channels
asgi_redis